refactor: move all Lore specific modules to new `lore` lib module
[What]

Create a module named `lore` in the library crate to house all
domain-specific modules. This considerably cleans the `src/` directory,
with the only con of adding another nesting level to the library crate
and having to collaterally update a handful of imports.

[Why/Context]

Currently, the `src/` directory of the project is a mess as it is
crowded with the modules defined in the library crate and the binary
crate as well. Specifically, the modules from the lib crate (defined in
`lib.rs`) are domain-specific, i.e., they represent implementations that
are significantly or entirely coupled with the Lore objects and API.

In this sense, we could extract every module from the library crate (at
the moment, they all live directly in `src/`) to a directory `lib/`.
This would result in a good clean-up, as all directories of these
modules, like `lore_api_client/`, `lore_session/`, etc., would be
encapsulated inside `lib/`, but the modules themselves
(`lore_api_client.rs`, `lore_session.rs`), would still need to be
directly under `src/`.
