+++ src/app/screens/bookmarked.rs
@@ -1,4 +1,4 @@
-use patch_hub::patch::Patch;
+use patch_hub::lore::patch::Patch;
 
 pub struct BookmarkedPatchsetsState {
     pub bookmarked_patchsets: Vec<Patch>,
