+++ src/lib.rs
@@ -1,4 +1,1 @@
-pub mod lore_api_client;
-pub mod lore_session;
-pub mod mailing_list;
-pub mod patch;
+pub mod lore;
