+++ src/lore.rs
@@ -1,0 +1,4 @@
+pub mod lore_api_client;
+pub mod lore_session;
+pub mod mailing_list;
+pub mod patch;
