+++ src/lore/mailing_list.rs
