+++ src/lore/patch/tests.rs
